refactor(Point): 패키지 이동
