+++ src/main/java/domain/util/PointGenerator.java
@@ -1,5 +1,7 @@
 package domain.util;
 
+import domain.ladder.Point;
+
 public interface PointGenerator {
 	Point generate();
 }
