+++ src/main/java/domain/util/RandomPointGenerator.java
@@ -2,6 +2,8 @@ package domain.util;
 
 import java.util.Random;
 
+import domain.ladder.Point;
+
 public class RandomPointGenerator implements PointGenerator {
 	@Override
 	public Point generate() {
