+++ src/test/java/domain/ladder/LadderTest.java
@@ -9,7 +9,6 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import domain.util.Point;
 import domain.util.PointGenerator;
 
 class LadderTest {
