+++ src/test/java/domain/util/LinePointsGeneratorTest.java
@@ -6,6 +6,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import domain.ladder.Point;
+
 class LinePointsGeneratorTest {
 
 	@Test
